Integrating importable_info into thesaurus
