--- lib/mumuki/bridge.rb
@@ -1,1 +1,0 @@
-require_relative './bridge/runner'
