+++ lib/mumuki/domain.rb
@@ -50,6 +50,5 @@ class Mumukit::Randomizer
   end
 end
 
-require_relative './bridge'
 require_relative './platform'
 require_relative './sync'
